BottomNavigationBar with Pagination
