+++ GameReview/lib/Screens/HomePages/UserProfilePage.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class UserProfile extends StatefulWidget {
+  const UserProfile({super.key});
+
+  @override
+  State<UserProfile> createState() => _UserProfileState();
+}
+
+class _UserProfileState extends State<UserProfile> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Text("Argane Mohamed"),
+      ),
+    );
+  }
+}
